+++ actix-router/src/router.rs
@@ -29,7 +29,7 @@ impl<T, U> Router<T, U> {
         profile_method!(recognize);
 
         for item in self.0.iter() {
-            if item.0.match_path(resource.resource_path()) {
+            if item.0.is_path_match(resource.resource_path()) {
                 return Some((&item.1, ResourceId(item.0.id())));
             }
         }
@@ -44,7 +44,7 @@ impl<T, U> Router<T, U> {
         profile_method!(recognize_mut);
 
         for item in self.0.iter_mut() {
-            if item.0.match_path(resource.resource_path()) {
+            if item.0.is_path_match(resource.resource_path()) {
                 return Some((&mut item.1, ResourceId(item.0.id())));
             }
         }
@@ -64,7 +64,7 @@ impl<T, U> Router<T, U> {
         profile_method!(recognize_checked);
 
         for item in self.0.iter() {
-            if item.0.match_path_checked(resource, &check, &item.2) {
+            if item.0.is_path_match_fn(resource, &check, &item.2) {
                 return Some((&item.1, ResourceId(item.0.id())));
             }
         }
@@ -84,7 +84,7 @@ impl<T, U> Router<T, U> {
         profile_method!(recognize_mut_checked);
 
         for item in self.0.iter_mut() {
-            if item.0.match_path_checked(resource, &check, &item.2) {
+            if item.0.is_path_match_fn(resource, &check, &item.2) {
                 return Some((&mut item.1, ResourceId(item.0.id())));
             }
         }